refactor: decouple convention grab-bag (Naming/Builtins to ir, dafny to its own module)
The `convention` module had accreted components unrelated to REST/SQL conventions
that other layers reached into. Re-layer them:

- `Naming` (generic case-conversion; used by testgen 14x / codegen / profile) and
  `Builtins` (the spec-language builtin registry) move to `ir` (the shared foundation).
- the Dafny code generator (`convention/dafny/`) becomes its own `dafny` module
  (main deps: `ir` only; golden tests reuse convention's SpecFixtures in test scope).

Consequently `lint` and `synth`-main no longer depend on `convention` at all:
- lint used convention only for `Builtins` (now in ir);
- synth-main used convention only for `convention.dafny.DafnyMethodHeader` (now `dafny`).

`Schema`/`Classify` deliberately stay in `convention` — moving them would create a
`profile`/`codegen` dependency cycle.

No proofs or generated code touched (no Isabelle regen). Full suite green including
the Dafny golden tests (byte-identical output); scalafmt + scalafix clean.
