+++ modules/cli/src/main/scala/specrest/cli/Inspect.scala
@@ -6,8 +6,8 @@ import io.circe.Json
 import io.circe.Printer
 import io.circe.syntax.EncoderOps
 import specrest.convention.Classify
-import specrest.convention.dafny.DafnyMethodHeader
-import specrest.convention.dafny.Generator as DafnyGenerator
+import specrest.dafny.DafnyMethodHeader
+import specrest.dafny.Generator as DafnyGenerator
 import specrest.ir.Serialize.given
 import specrest.ir.VerifyError
 import specrest.ir.generated.SpecRestGenerated.*
