+++ modules/codegen/src/main/scala/specrest/codegen/python/EmitPython.scala
@@ -24,7 +24,7 @@ import specrest.codegen.migration.SchemaDiff
 import specrest.codegen.migration.SchemaSnapshot
 import specrest.codegen.openapi.OpenApi
 import specrest.convention.EndpointSpec
-import specrest.convention.Naming
+import specrest.ir.Naming
 import specrest.ir.generated.SpecRestGenerated.*
 import specrest.profile.ProfiledEntity
 import specrest.profile.ProfiledField
