+++ modules/convention/src/main/scala/specrest/convention/Schema.scala
@@ -1,5 +1,5 @@
 package specrest.convention
-
+import specrest.ir.Naming
 import specrest.ir.generated.SpecRestGenerated.*
 import specrest.ir.idx
