+++ modules/dafny/src/test/scala/specrest/dafny/GeneratorGoldenTest.scala
@@ -1,4 +1,4 @@
-package specrest.convention.dafny
+package specrest.dafny
 
 import cats.effect.IO
 import munit.CatsEffectSuite
