+++ modules/ir/src/main/scala/specrest/ir/Naming.scala
@@ -1,4 +1,4 @@
-package specrest.convention
+package specrest.ir
 
 object Naming:
