+++ modules/testgen/src/main/scala/specrest/testgen/AdminRouterTs.scala
@@ -1,6 +1,6 @@
 package specrest.testgen
 
-import specrest.convention.Naming
+import specrest.ir.Naming
 import specrest.ir.generated.SpecRestGenerated.entFields
 import specrest.ir.generated.SpecRestGenerated.entName
 import specrest.ir.generated.SpecRestGenerated.entity_decl_full
