+++ modules/testgen/src/main/scala/specrest/testgen/Templates.scala
@@ -1,6 +1,6 @@
 package specrest.testgen
 
-import specrest.convention.Naming
+import specrest.ir.Naming
 import specrest.ir.generated.SpecRestGenerated.*
 
 import java.io.ByteArrayOutputStream
